predictions-docker/app/requirements.txt
transformers==4.30.2
torch==1.12.1
flask
google-cloud-storage